requirements.txt
Flask>=2.0
flask-socketio>=5.0
python-socketio>=5.0
openai>=1.0.0
eventlet>=0.40.3
python-dotenv>=1.0.0